Add Java7 experimental traverser
Simple file system traverser which uses Java7 classes; extracts ACL
information from Linux CIFS implementation and decodes it; uses jCIFS
to resolve SIDs.

Code is not made to be pretty, just to test that things work and to
measure performance and reliability of the solution. We can later
re-use some of the pieces. Most options are hard-coded.