Adding MockFileDelegate and MockFile
The purpose of these classes is to allow construction of
in-memory file system structure with configuration of the
files and directories to suit the tests that will be performed.

MockFile represents a file or directory.  It has many setters
that allow the tests to define the file's properties and
configuration.

MockFileDelegate is a FileDelegate implementation over the
in-memory file system consisting of MockFiles.

Code Review: https://codereview.appspot.com/47730043/