Updating newDocId to follow links.
It's prefered that newDocId follow links to since, if some reason, the
path is a link we will generate an ID for the linked target file. This
would happen if the file was deleted and a link created after an initial
crawl. In this instance, newDocId would generate a DocId that is different
from the DocId already sent to the GSA which would cause the FS adaptor to
log a message and skip this file.