Add WindowsFileDelegateTest
This adds a basic WindowsFileDelegateTest.  It tests newDocId() and
most file system monitor functions using the local file system.
It only runs on Windows.

This change adds a startSignal CountDownLatch that the monitor uses
to signal that the monitor is actively listening for changes, and
startMonitorPath() waits for the monitor thread to be up and running
before returning.

TODO: Add tests for ACL features.
TODO: Add tests for DFS features.

Code Review: https://codereview.appspot.com/57170048/