Add WindowsAclFileAttributeViewsTest (Part 3)
This final CL adds tests for the getShareAcl() method.  I was forced
to make WindowsAclFileAttributeViews.Mpr class public and static or
jna.Native complained when I tried to create an annonymous instance of it.
This CL also required adding another stub JNA implemenation of Netapi32.

This brings coverage for WindowsAclFileAttributeViews up to
96% line coverage and 89% branch coverage.  It brings overall
coverage for the adaptor to 84% line coverage and 80% branch
coverage.

Code Review: http://codereview.appspot.com/83900043