Adding use of AsyncDocIdPusher
Adding changes to allow the adaptor so use the AsyncDocIdPusher interface. With
using AsyncDocIdPusher, the adaptor will not need the monitor thread pushes the
changes to the GSA. Instead, when an file incremental change notification is
sent to the adaptor, the adaptor will pushed file Records to the
AsyncDocIdPusher. The library will be managing batching and send the Records to
the GSA.