Improving handling of last access time
Updating the adaptor to call the Win Api SetFileTime with the value of -1 for
the last access time before reading the content of a file. This will stop
Windows from updating the last access time as the file content is read by the
adaptor. As a fail over, the adaptor also queries the last access time of a
document before reading the content and then check if the time has changed. If
it has, then the adaptor will attempt to reset the last access time of the
document back to the original.