Fix WindowsAclViews test failures
There are some test failures on the 32-bit platforms.
It seems the tests were having some issues serialzing and deserialzing
SIDs correctly. This does not appear to be a 32-bit issue, so much
as a GC issue. It looks like the SIDs were getting garbage collected
because the only reference to them was a serialized Pointer in the
ACE.

This change adds caching of SIDs in a HashMap to the tests to avoid
serializing and deserializing of SIDs.

Code Review: https://codereview.appspot.com/132800043/