Fix b/12897821: Enumerate Namespace target links.
This change allows the administrator to specify a DFS Namespace
(aka DFS Root), as the startpoint for the filesystem adaptor.
If this is done, the adaptor will enumerate the DFS links in
that namespace at crawl time, and supply them as links in a
generated HTML document (similar to how directories were already
handled).

The DFS links are resolved when the GSA crawls them, the share ACLs
for the link are returned in the response, and a new filesystem
monitor is started for the target filesystem (if one has not
already running).

The two biggest structual changes relate to the multiple targets
being traversed:

1) There can now be multiple filesystem monitors running.

2) The names of the Share ACL and DFS Share ACL have changed.
   In the past, they were naked fragment names.  Now they are
   fragments on the filesystem root or DFS link DocIds.
   This change is incompatible with existing indexed content,
   and some or all documents may require a recrawl.

Code Review: http://codereview.appspot.com/184950043