Refactor FsAdaptor share ACL handling to support multiple shares
In order to support the changes for the following bugs:

b/6513820: Enumerate share points or Namespaces given a hostname
b/12897821: Enumerate Namespace target links

Supplying only hostname or DFS Namespace as a startpoint
could eventually resolve to multiple shares being traversed.
In anticipation of these changes, this refactoring alters the
FsAdaptor code that reads and pushes DFS share ACLs and share ACLs
to not assume the rootPath is the share.

Code Review: http://codereview.appspot.com/180940043