Rename DfsRoot to DfsNamespace to avoid confusion.
Even though MSDN developer documentation refers to a DFS Namespace as a
DFS Root, we have multiple uses of the work "root" in the adaptor.  So
to avoid some confusion, we have decided to unambiguously refer to a
DFS Namespace as DfsNamespace, rather than DfsRoot.

Code Review: http://codereview.appspot.com/187280043