Support multiple start paths in filesystemadaptor.src config property.
This change allows the adaptor to accept multiple start paths as the
configured file system sources. The multiple sources are specified as
a delimited list of UNC paths for the filesystemadaptor.src property.

The default delimiter is semicolon, but is configurable with the
filesystemadaptor.src.separator property.
If the filesystemadaptor.src.separator property is empty string,
then the filesystemadaptor.src property is considered to be a single
path (probably chock full of whitespace and punctuation characters).

Code Review: https://codereview.appspot.com/194730044/