Change order of validity checks in getDocContent()
This change demotes the isVisibleDescendantOfRoot check to the
bottom. It was previously done just after validating the docid was
one of ours, followed by checks for existence and isFileOrFolder.

isVisibleDescendantOfRoot() is computationally expensive, but
the most interesting feedback from testing deals with the GSA
crawling relative hyperlinks embedded in the crawled files.
If the files containing relative hyperlinks are not in their
deployment configuration, the resolved URLs point to files that
do not exist.

This change promotes existence check above all others except
valid docid.  More importantly, this eliminates "Skipping <file>
because it is not a decendant of a start path" warnings where
resolving a relative hyperlink backs up past the start point.

Code Review: http://codereview.appspot.com/198760043