Fix b/19036758: Adaptor fails to start if one of multiple startPaths is invalid
This change handles IOExceptions in the loop that validates startPaths.
This allows the adaptor to proceed with valid startPaths, even if one
or more are broken. I added more granular exception handling in
validateShare(), to catch conditions that are definitely configuration
errors. However it is difficult to impossible to differentiate between
a misspelled DFS namespace, DFS link, or UNC share and one that is
off-line.

Code Review: http://codereview.appspot.com/199750043