Add StatusSource Support
This changes adds a Dashboard Status source for each start path
and/or file system root. An initial status is set during the
validateStartPaths() portion of init(), and updated during crawl
requests.

There is also a periodic statusUpdaterService that checks on
the status of all the sources.  A new configuration parameter,
filesystemadaptor.statusUpdateInterval, specifies in minutes
how often to refresh the status.  The default interval is 15 minutes.

Code Review: http://codereview.appspot.com/198870043