Remove DocId validity check from FsAdaptor.getDocContent()
This change removes a DocId validity check from getDocContent()
that attempts to verify that the recieved DocId was generated
by this adaptor. Yet, the adaptor had already created a Path
from the docid, and further checks will verify it is a Path
under our start points.  So I feel this check was superfluous.

Profiling showed the adaptor was often spending 20%-50% of
its time in newDocId().

Code Review: https://codereview.appspot.com/196180043/