Fix b/14982270: Detect inability to restore last access time
This change detects the inability to restore the last access time
of crawled files and folders and rejects further crawl requests
for that filesystem.

Failure to preserve last access times can fool backup and archive systems
into thinking the file or folder has been recently accessed by a human,
preventing the movement of least recently used items to secondary storage.

If the adaptor is unable to restore the last access time for the file,
it is likely the traversal user does not have sufficient privileges to
write the file's attributes.  We therefore halt crawls on this volume
to prevent altering the last access timestamps for potentially thousands
of files.

A new  configuration parameter, filesystemadaptor.preserveLastAccessTime,
controls the enforcement of preservation of the last access timestamps.
It may have one of 3 values:
  ALWAYS: The adaptor will attempt to preserve the last access time
          for all files and folders crawled. The first failure to do so
          will force the adaptor to reject all subsequent crawl requests
          for the filesystem to prevent altering the last access timestamps
          for potentially thousands of files.

  IF_ALLOWED: The adaptor will attempt to preserve the last access time
              for all files and folders crawled, even though some file
              timestamps might not be preserved.

  NEVER: The adaptor will make no attempt to preserve the last access
         time for crawled files and folders.

The default value is ALWAYS.

Code Review: http://codereview.appspot.com/221760043