Fix inaccurate log message in isVisibleDescendantOfRoot
This change fixes a subtle error in isVisibleDescendantOfRoot().
If a non-hidden file was the immediate child of a hidden folder,
the code would incorrectly log that the file itself was hidden,
rather than it was hidden under its parent. The returned value
from isVisibleDescendantOfRoot() is correct, but it is the log
message that is misleading.

Code Review: http://codereview.appspot.com/246690043