Fix b/22194167: Make log message strings uniquely identify spots in source.
This change makes sure no two log messages are identical, so that
they may be associated with a particular location in the source code.
It also cleans up some of the access-controlled serving log messages
to the conventions used by the rest of the adaptor.

Code review: http://codereview.appspot.com/246680043/