Catch Not Found exceptions when reading file and directory content.
This change catches and logs FileNotFoundException and NoSuchFileException
when reading directory and file content.

While examining a customer's logs, NoSuchFileException was being thrown
out of Files.newDirectoryStream() when crawling a bunch of directories.
I found this odd because this happens near the end of getDocContent().
We have successfully read the attributes, metadata, and ACL by that point,
and isDirectory() returns true; but apparently those were all phantoms.

It is not clear to me what could cause such a condition.  It could be
caching on the part of DFS. It could be an off-line file or filesystem.
It could be a corrupt filesystem where we can read the directory, but
the inodes for the content are missing. It could be a permissions problem
where the traversal user does not have list files permission (although
I would expect AccessDenied).

Code review: http://codereview.appspot.com/260280043