Fix b/14447323: Add support for non-root startpaths
This change adds support for startpaths that are not at the
root of a file share, a frequent customer request. The deeper
startpaths supply an ACL that includes permissions inherited
from it ancestors.

This change revealed a flaw when testing ACLs on DFS paths
which prompted two changes:
 * ACLs on verious nodes now use unique group names to
   make sure the correct ACL is used in the named resource.
 * I dispensed with the sketchy MockFileDelegate.getFileOrRoot()
   slight-of-hand when dealing with DFS and Share ACLs.
   The targets of DFS links now have a separate root in the
   mock filesystem.

Code Review: http://codereview.appspot.com/271210043