Fix b/24719464: Use hybrid response for large directory listings
Large directories can have tens or hundreds of thousands of files.
The GSA truncates large HTML documents at 2.5MB, so we return
the first maxHtmlLinks worth as HTML content and the rest as external
anchors.  This code was adapted from the hybrid mechanism used by
the Documentum Adaptor.

This change adds a new configuration property:
filesystemadaptor.maxHtmlSize, whose default value is 1000 links.
The property name is "Size", rather than "Links" to be consistent
with the property name in the Sharepoint and Documentum adaptors.

Code Review: http://codereview.appspot.com/269250043