Fix b/25749810: Support If-Modified-Since with 204 responses
This change adds support for If-Modified-Since to the adaptor,
allowing it to return no content if the file/folder has not been
modified since it was last crawled.  ACLs and Metadata are still
returned, but no content is returned.

It also simplifies the MockRequest structure somewhat, making it
even more trivial.

Code Review: http://codereview.appspot.com/275390043