Workaround b/26441707: GSA fails to crawl path containing semicolon
The GSA has a bug where it fails to properly canonicalize relative
URLs that contain a semicolon. As a result, the resolved URL is
missing the path component which contains the semicolon, and the
subsequent crawl request returns a 404.

This change patches HtmlResponseWriter so if the path contains a semicolon,
force the HTML link href to be the full URL, rather than a relative URL.

Note: The GSA has a separate bug that incorrectly displays documents
whose URL contain a semicolon in Index Diagnostics. However, that
bug does not impact the ability to index or search the documents.
This work-around has no effect on that separate Index Diagnostics bug.

Code Review: http://codereview.appspot.com/289310043