Fix b/27579240: Feed large directory listings asynchronously to avoid timeouts.
This change returns at most maxHtmlSize links in the generated HTML
of a directory listing. If the directory has more than maxHtmlSize
entries, the directory stream is closed and the generated HTML is
returned as the response. A separate thread will then be launched
to push the directory entries via the DocIdPusher. This should avoid
getDocContent timeouts reading large directories.