Fix b/28133136: Change notifications cease after file server reboot
If the file server gets rebooted, the connector stops receiving
change notifictions. Any IOExceptions thrown out of the monitor
would kill the monitor thread. This change handles IOExceptions
in the thread, and retrys the monitor with an exponential backoff.

This change also adds a 15 minute timeout to the monitor thread
when waiting for change notifications from the filesystem, in case
the server off-line goes unnoticed while waiting.